Revert "Move typing.py -> _types.py"
This reverts commit 23069a59d26f3b9ed1ee7aea39812687d7098e43.
