+++ mne_bids_pipeline/_config.py
@@ -7,7 +7,7 @@ from annotated_types import Ge, Interval, Len
 from mne import Covariance
 from mne_bids import BIDSPath
 
-from mne_bids_pipeline._types import (
+from mne_bids_pipeline.typing import (
     ArbitraryContrast,
     DigMontageType,
     FloatArrayLike,
